+++ dallinger/docker/prepare_docker_image.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+# If you need some extra requirement available to experiment code
+# (for example the ffmpeg binary) you can add a prepare_docker_image.sh
+# file to your experiment directory.
+# It will be run when preparing a docker image. Anything installed
+# by this script will be available to the experiment code.
+
+# This experiment did not specify any prepare_docker_image.sh script,
+# so this empty template was used.